# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Rafiq Al-Quran API
  version: 2.19.0
  description: Release baseline API contract for current backend endpoints.
servers:
  - url: http://localhost:4000
    description: Current runtime
  - url: http://localhost:4000
    description: Development
  - url: https://staging.api.rafiq.example
    description: Staging
  - url: https://api.rafiq.example
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Session payload
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthSessionResponse"
        "400":
          description: Invalid body
          content:
            application/json:
              schema: *a1
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: *a1
        "429":
          description: Rate limited
          content:
            application/json:
              schema: *a1
  /auth/forgot-password:
    post:
      tags:
        - Auth
      summary: Request password reset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Password reset request accepted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/GenericMessageResponse"
        "400":
          description: Invalid body
          content:
            application/json:
              schema: *a1
        "429":
          description: Rate limited
          content:
            application/json:
              schema: *a1
  /auth/reset-password:
    post:
      tags:
        - Auth
      summary: Reset password with token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/GenericMessageResponse"
        "400":
          description: Invalid token or body
          content:
            application/json:
              schema: *a1
        "429":
          description: Rate limited
          content:
            application/json:
              schema: *a1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessEnvelope:
      type: object
      properties:
        ok:
          type: boolean
          enum:
            - true
        data:
          type: object
          additionalProperties: true
      required:
        - ok
...
    Role:
      type: string
      enum:
        - SUPER_ADMIN
        - CENTER_ADMIN
        - SUPERVISOR
        - TEACHER
        - PARENT
        - STUDENT
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/Role"
...
    LoginRequest:
      type: object
      required:
        - identifier
        - password
      properties:
        identifier:
          type: string
          description: Email or phone number.
        email:
          type: string
          description: "Deprecated: legacy alias for identifier."
          deprecated: true
        password:
          type: string
          minLength: 8
    ForgotPasswordRequest:
      type: object
      required:
        - identifier
      properties:
        identifier:
          type: string
          description: Email or phone number.
    ResetPasswordRequest:
      type: object
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
          minLength: 20
          maxLength: 4000
        newPassword:
          type: string
          minLength: 8
          maxLength: 128
    GenericMessageResponse:
      type: object
      properties:
        message:
          type: string
...
    AuthSessionResponse:
      type: object
      properties:
        accessToken:
          type: string
        accessExpiresIn:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
...